back/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
asttokens==3.0.0
    # via stack-data
attrs==24.3.0
    # via
    #   jsonschema
    #   referencing
build==1.2.2.post1
    # via pip-tools
click==8.1.8
    # via pip-tools
decorator==5.1.1
    # via ipython
django==5.1.2
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-cors-headers==4.6.0
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
drf-spectacular==0.28.0
    # via -r requirements.in
executing==2.1.0
    # via stack-data
faker==30.6.0
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
ipython==8.31.0
    # via -r requirements.in
jedi==0.19.2
    # via ipython
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
matplotlib-inline==0.1.7
    # via ipython
packaging==24.2
    # via build
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pip-tools==7.4.1
    # via -r requirements.in
prompt-toolkit==3.0.48
    # via ipython
psycopg2-binary==2.9.10
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pygments==2.18.0
    # via ipython
pyjwt==2.9.0
    # via djangorestframework-simplejwt
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via faker
pyyaml==6.0.2
    # via drf-spectacular
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
ruff==0.8.5
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sqlparse==0.5.1
    # via django
stack-data==0.6.3
    # via ipython
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.12.2
    # via faker
uritemplate==4.1.1
    # via drf-spectacular
wcwidth==0.2.13
    # via prompt-toolkit
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
